union review's operations, rebase Product and DetailedProduct models
