+++ GBShop/GBShop/App/AppDelegate.swift
@@ -43,14 +43,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     private func regRequest() {
         let reg = requestFactory.makeRegistrationRequestFactory()
 
-        reg.registration(
+        let profile = SignUpUser(
             username: "Somebody",
             password: "mypassword",
             email: "some@some.ru",
-            gender: "m",
-            creditCardNumber: "9872389-2424-234224-234",
-            aboutMe: "This is good! I think I will switch to another language"
-        ) { response in
+            creditCard: "9872389-2424-234224-234",
+            gender: .man,
+            bio: "This is good! I think I will switch to another language")
+
+        reg.registration(profile: profile) { response in
             switch response.result {
             case .success(let message):
                 print(message)
@@ -80,7 +81,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     private func editRequest() {
         let edit = requestFactory.makeEditProfileRequestFactory()
         
-        let editProfileUserModel = EditProfileUser(
+        let profile = EditUserProfile(
             id: 123,
             username: "Somebody",
             password: "mypassword",
@@ -89,7 +90,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             gender: .man,
             bio: "This is good! I think I will switch to another language")
 
-        edit.editProfile(editProfileUser: editProfileUserModel) { response in
+        edit.editProfile(profile: profile) { response in
             switch response.result {
             case .success(let login):
                 print(login)
@@ -162,9 +163,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     // MARK: - Add product review
 
     private func addReview() {
-        let getReviews = requestFactory.makeAddReviewRequestFactory()
+        let addReview = requestFactory.makeReviewsRequestFactory()
 
-        getReviews.addReview(userId: 123, productId: 456, description: "Текст отзыва") { response in
+        addReview.addReview(userId: 123, productId: 456, description: "Текст отзыва") { response in
             switch response.result {
             case .success(let response):
                 print(response)
@@ -173,7 +174,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             }
         }
 
-        getReviews.addReview(userId: nil, productId: 456, description: "Текст отзыва") { response in
+        addReview.addReview(userId: nil, productId: 456, description: "Текст отзыва") { response in
             switch response.result {
             case .success(let response):
                 print(response)
@@ -186,9 +187,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     // MARK: - Approve product review
 
     private func approveReview() {
-        let getReviews = requestFactory.makeApproveReviewRequestFactory()
+        let approveReview = requestFactory.makeReviewsRequestFactory()
 
-        getReviews.approveReview(userId: 123, reviewId: 112) { response in
+        approveReview.approveReview(userId: 123, reviewId: 112) { response in
             switch response.result {
             case .success(let response):
                 print(response)
@@ -201,9 +202,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     // MARK: - Remove product review
 
     private func removeReview() {
-        let getReviews = requestFactory.makeRemoveReviewRequestFactory()
+        let removeReview = requestFactory.makeReviewsRequestFactory()
 
-        getReviews.removeReview(userId: 123, reviewId: 112) { response in
+        removeReview.removeReview(userId: 123, reviewId: 112) { response in
             switch response.result {
             case .success(let response):
                 print(response)
