--- GBShop/GBShop/App/BusinessLogic/Network/ApproveReview/ApproveReviewRequestFactory.swift
@@ -1,19 +1,0 @@
-//
-//  ApproveReviewRequestFactory.swift
-//  GBShop
-//
-//  Created by Artem Mayer on 27.02.2023.
-//
-
-import Alamofire
-
-protocol ApproveReviewRequestFactory {
-
-    // MARK: - Functions
-
-    func approveReview(
-        userId: Int,
-        reviewId: Int,
-        completionHandler: @escaping (AFDataResponse<ApproveReviewResult>) -> Void)
-}
-
