+++ GBShop/GBShop/App/BusinessLogic/Network/Reviews/ReviewsRequestFactory.swift
@@ -15,4 +15,20 @@ protocol ReviewsRequestFactory {
         productId: Int,
         pageNumber: Int,
         completionHandler: @escaping (AFDataResponse<ReviewsResult>) -> Void)
+
+    func addReview(
+        userId: Int?,
+        productId: Int,
+        description: String,
+        completionHandler: @escaping (AFDataResponse<AddReviewResult>) -> Void)
+
+    func approveReview(
+        userId: Int,
+        reviewId: Int,
+        completionHandler: @escaping (AFDataResponse<ApproveReviewResult>) -> Void)
+
+    func removeReview(
+        userId: Int,
+        reviewId: Int,
+        completionHandler: @escaping (AFDataResponse<RemoveReviewResult>) -> Void)
 }
