+++ GBShop/GBShop/App/Entity/ResultsEntity/CatalogResult.swift
@@ -7,27 +7,6 @@
 
 import Foundation
 
-// MARK: - Product
-
-struct Product: Codable {
-
-    // MARK: - CodingKeys
-
-    enum CodingKeys: String, CodingKey {
-        case id = "product_id"
-        case name = "product_name"
-        case price = "product_price"
-    }
-
-    // MARK: - Properties
-
-    let id: Int
-    let name: String
-    let price: Int
-}
-
-// MARK: - CatalogResult
-
 struct CatalogResult: Codable {
 
     // MARK: - CodingKeys
