+++ GBShop/GBShop/App/Entity/ResultsEntity/ProductResult.swift
@@ -7,23 +7,6 @@
 
 import Foundation
 
-struct ProductDetailed: Codable {
-
-    // MARK: - CodingKeys
-
-    enum CodingKeys: String, CodingKey {
-        case price = "product_price"
-        case name = "product_name"
-        case description = "product_description"
-    }
-
-    // MARK: - Properties
-
-    let price: Int
-    let name: String
-    let description: String
-}
-
 struct ProductResult: Codable {
 
     // MARK: - CodingKeys
@@ -36,5 +19,5 @@ struct ProductResult: Codable {
     // MARK: - Properties
 
     let result: Int
-    let product: ProductDetailed
+    let product: DetailedProduct?
 }
